LayoutTests/imported/w3c/web-platform-tests/html/interaction/focus/focusgroup/tentative/forward-navigation/wraps-successfully-expected.txt

item1 item2

FAIL When the focus is set on the last item of a focusgroup that supports wrapping in the axis of the arrow key pressed, the focus should move back to the first item. assert_equals: expected Element node <span id="item1" tabindex="0">item1</span> but got Element node <span id="item2" tabindex="-1">item2</span>
